% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-layout.R
\name{animate_layout}
\alias{animate_layout}
\title{Animate a funnel in a pedigreeburst}
\usage{
animate_layout(funnels, funnel)
}
\arguments{
\item{funnels}{A data.frame}

\item{funnel}{A string describing the funnel of interest}
}
\value{
A gganim object
}
\description{
Animate a funnel in a pedigreeburst
}
