% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare data for pedigree burst}
\usage{
data_prep(funnels, padding = 0.1)
}
\arguments{
\item{funnels}{A data.frame}

\item{padding}{Numeric. How much space between crosses. Default 0.1}
}
\value{
A data.frame
}
\description{
Prepare data for pedigree burst
}
