% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel-layout.R
\name{funnel_layout}
\alias{funnel_layout}
\title{Prepare data for pedigree burst}
\usage{
funnel_layout(funnels, focus_level = 1, padding = 0.1)
}
\arguments{
\item{funnels}{A data.frame}

\item{padding}{Numeric. How much space between crosses. Default 0.1}
}
\value{
A data.frame
}
\description{
Prepare data for pedigree burst
}
