% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel-to-pedigree.R
\name{funnel_to_pedigree}
\alias{funnel_to_pedigree}
\title{Funnel to pedigree}
\usage{
funnel_to_pedigree(funnels, include_zero = FALSE)
}
\arguments{
\item{funnels}{A dataframe describing funnels}

\item{include_zero}{Logical. Whether to include the zero (non-crossing) generation}
}
\value{
A dataframe
}
\description{
Funnel to pedigree
}
