% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnelburst.R
\name{funnelburst}
\alias{funnelburst}
\title{Funnelburst}
\usage{
funnelburst(funnels, focus_level = 1, show_levels = 1:8,
  print_founders = TRUE, show_legend = FALSE, rotate_labels = TRUE,
  xmax = NULL, colour_by_founder = TRUE, padding = 0.1,
  polar = TRUE)
}
\arguments{
\item{funnels}{A data.frame}

\item{show_levels}{A numeric vector describing which levels to show in the figure. Must start at 1}

\item{print_founders}{Logical. Should founder names be printed in the figure?}

\item{show_legend}{Logical. Should a legend be printed?}

\item{rotate_labels}{= TRUE}

\item{xmax}{numeric. Controls limits if a subset of funnels are used}

\item{colour_by_founder}{Logical. Default TRUE. If FALSE a grey figure is returned}

\item{padding}{numeric. Controls spacing between crossing levels}
}
\value{
a ggplot2 object
}
\description{
Funnelburst
}
