% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-graph-layout.R
\name{pedigree_graph_layout}
\alias{pedigree_graph_layout}
\title{Pedigree graph layout}
\usage{
pedigree_graph_layout(pedigree)
}
\arguments{
\item{pedigree}{A pedigree dataframe with columns: f, m, id, gen}
}
\value{
A dataframe
}
\description{
Pedigree graph layout
}
