% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeburst.R
\name{pedigreeburst}
\alias{pedigreeburst}
\title{Pedigreeburst}
\usage{
pedigreeburst(funnels, show_levels = 1:8, print_founders = TRUE,
  show_legend = FALSE, rotate_labels = TRUE, xmax = NULL,
  colour_by_founder = TRUE, padding = 0.1)
}
\arguments{
\item{funnels}{A data.frame}

\item{show_levels}{A numeric vector describing which levels to show in the figure. Must start at 1}

\item{print_founders}{Logical. Should founder names be printed in the figure?}

\item{show_legend}{Logical. Should a legend be printed?}

\item{rotate_labels}{= TRUE}

\item{xmax}{numeric. Controls limits if a subset of funnels are used}

\item{colour_by_founder}{Logical. Default TRUE. If FALSE a grey figure is returned}

\item{padding}{numeric. Controls spacing between crossing levels}
}
\value{
a ggplot2 object
}
\description{
Pedigreeburst
}
