% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-funnel.R, R/animate-layout.R,
%   R/funnel-layout.R, R/funnelburst.R, R/pedigree-graph.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aes}
\alias{mutate}
\alias{.data}
\alias{case_when}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{aes}}}

  \item{rlang}{\code{\link[rlang]{.data}}}
}}

