% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_num}
\alias{str_to_num}
\title{Convert character string to numbers}
\usage{
str_to_num(string)
}
\arguments{
\item{string}{}
}
\value{
character vector
}
\description{
Convert character string to numbers
}
\examples{
str_to_num("AEHD")
}
