% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_graph_layout.R
\name{pedigree_graph_layout}
\alias{pedigree_graph_layout}
\title{Pedigree graph layout}
\usage{
pedigree_graph_layout(funnels)
}
\arguments{
\item{funnels}{A data frame with funnels}
}
\description{
Pedigree graph layout
}
